Improve the SlimefunReloadedItem framework. All component based (#6)
This also improves the machine code and moves it to a state-based system. Note that this commit does not include the registration of states or the invocation of the methods defined in the machine nor item components. This will be done at a later date once the planned methods have been finalized. Until such time, implementations of these classes may be created without much required change in the future.